// src/shared/proto/custom_enums/v1/custom_enums.ts
export enum OperationStatus {
  /** OPERATION_STATUS_UNSPECIFIED - Дефолтный (неопределенный) статус операции */
  OPERATION_STATUS_UNSPECIFIED = "OPERATION_STATUS_UNSPECIFIED",
  /**
   * OPERATION_STATUS_OK - Статус операции сигнализирующий о корректом завершении операции
   *  http analog status code: 200
   *  gRPC analog status code: 1
   */
  OPERATION_STATUS_OK = "OPERATION_STATUS_OK",
  /**
   * OPERATION_STATUS_ERROR - Статус операции сигнализирующий о не корректом завершении операции
   *  http analog status code: 400
   *  gRPC analog status code: 8,9,11
   */
  OPERATION_STATUS_ERROR = "OPERATION_STATUS_ERROR",
  /**
   * OPERATION_STATUS_NOT_PERMITTED - Статус операции сигнализирующий о недостаточном наборе прав для его вызова
   *  http analog status code: 403
   *  gRPC analog status code: 7
   */
  OPERATION_STATUS_NOT_PERMITTED = "OPERATION_STATUS_NOT_PERMITTED",
  /**
   * OPERATION_STATUS_DUPLICATE - Статус операции сигнализирующий о том что операция продублирована
   *  http analog status code: 409
   *  gRPC analog status code: 6
   */
  OPERATION_STATUS_DUPLICATE = "OPERATION_STATUS_DUPLICATE",
  /**
   * OPERATION_STATUS_SERVER_ERROR - Статус операции сигнализирующий о не корректой обработке со стороны сервера
   *  http analog status code: 500
   *  gRPC analog status code: 13,14,15
   */
  OPERATION_STATUS_SERVER_ERROR = "OPERATION_STATUS_SERVER_ERROR",
  /**
   * OPERATION_STATUS_NO_DATA - Статус операции сигнализирующий о том, что данных по запросу не обнаружено
   *  http analog status code: 204
   *  gRPC analog status code: 5
   */
  OPERATION_STATUS_NO_DATA = "OPERATION_STATUS_NO_DATA",
  /** OPERATION_STATUS_DECLINED - Статус операции сигнализирующий о том, что при текущих условиях операцию выполнить нельзя */
  OPERATION_STATUS_DECLINED = "OPERATION_STATUS_DECLINED",
  /** OPERATION_STATUS_LOCKED - Статус операции - заблокировано */
  OPERATION_STATUS_LOCKED = "OPERATION_STATUS_LOCKED",
  UNRECOGNIZED = "UNRECOGNIZED",
}